+++ src/main.tsx
@@ -8,8 +8,8 @@ import CalendarMain from './pages/calendar/calendar-main';
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <ThemeProvider>
-      <CalendarMain />
       <App />
+      <CalendarMain />
     </ThemeProvider>
   </StrictMode>
 );
